// src/types.ts
export type FacilityType =
  | 'Clinic'
  | 'PHC'
  | 'CHC'
  | 'Sub-District Hospital'
  | 'District Hospital'
  | 'Specialist Medical College'
  | 'Tertiary Hospital'
  | 'Private Empanelled';

export interface Facility {
  id: string;
  name: string;
  type: FacilityType;
  category: 'Public' | 'Private' | 'Trust' | 'Associated';
  latitude: number;
  longitude: number;
  address: string;
  district: string;
  phone: string;
  emergencyCapability: boolean;
  totalBeds: number;
  occupiedBeds: number;
  icuBeds: number;
  occupiedIcuBeds: number;
  currentQueueLength: number;
  averageWaitTimeMin: number;
  specialists: { specialty: string; available: boolean; doctorName: string }[];
  services: string[];
  availableDiagnostics: string[];
  medicineStockRatio: number;
}